requirements.txt
numpy
pandas
streamlit
folium
streamlit_folium
pillow